+++ noodles-vcf/src/header/record/value/map/format.rs
@@ -120,7 +120,7 @@ impl fmt::Display for Map<Format> {
 
 impl From<&Key> for Map<Format> {
     fn from(key: &Key) -> Self {
-        let number = definition::number(key).unwrap_or(Number::Count(1));
+        let number = definition::number(key).unwrap_or_default();
         let ty = definition::ty(key).unwrap_or(Type::String);
         let description = definition::description(key)
             .map(|s| s.into())